# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rexxgo API",
        "description": "API documentation for Auth and Wallet modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Base server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "operationId": "9f8e027ac55656c09a892de8f0825e75",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile",
                "operationId": "8cf72731cd1a666d1008449e4636665f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "description": "Country code (USA or NG)",
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "NY"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "10001"
                                    },
                                    "ssn": {
                                        "description": "For US users only",
                                        "type": "string",
                                        "example": "123-45-6789"
                                    },
                                    "bvn": {
                                        "description": "For Nigeria users only",
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "nin": {
                                        "description": "For Nigeria users only",
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/kyc-status": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get KYC status",
                "operationId": "9786d0527e02cdd0d69e1d3ff2a81588",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}